Move TypeSelectorAssertion specs to own file
